Create add_user.py
This code initializes an instance of SimpleMFRC522 and prompts the user to hold an RFID tag near the reader. If an RFID tag is detected, the code takes a photo using the OpenCV library and saves it in the 'database' folder with the tag ID as the filename. The code waits for 5 seconds if an RFID tag is not detected and prompts the user again. The GPIO pins are cleaned up on keyboard interrupt.